+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -41,7 +41,6 @@ use Symfony\Component\PropertyAccess\PropertyAccessor;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractorInterface;
 use Symfony\Component\RateLimiter\Policy\TokenBucketLimiter;
 use Symfony\Component\Scheduler\Schedule;
-use Symfony\Component\Semaphore\Semaphore;
 use Symfony\Component\Serializer\Encoder\JsonDecode;
 use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\Translation\Translator;
@@ -179,7 +178,7 @@ class Configuration implements ConfigurationInterface
         $this->addExceptionsSection($rootNode);
         $this->addWebLinkSection($rootNode);
         $this->addLockSection($rootNode, $enableIfStandalone);
-        $this->addSemaphoreSection($rootNode, $enableIfStandalone);
+        $this->addSemaphoreSection($rootNode);
         $this->addMessengerSection($rootNode, $enableIfStandalone);
         $this->addSchedulerSection($rootNode, $enableIfStandalone);
         $this->addRobotsIndexSection($rootNode);
@@ -1426,59 +1425,15 @@ class Configuration implements ConfigurationInterface
         ;
     }
 
-    /**
-     * @param-immediately-invoked-callable $enableIfStandalone
-     */
-    private function addSemaphoreSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
+    private function addSemaphoreSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
-                ->arrayNode('semaphore')
-                    ->info('Semaphore configuration')
-                    ->acceptAndWrap(['string'], 'resources')
-                    ->{$enableIfStandalone('symfony/semaphore', Semaphore::class)}()
-                    ->beforeNormalization()
-                        ->ifArray()
-                        ->then(static function ($v) {
-                            if (!isset($v['resources']) && !isset($v['resource'])) {
-                                $v = ['resources' => $v];
-                                if (\array_key_exists('enabled', $v['resources'])) {
-                                    $v['enabled'] = $v['resources']['enabled'];
-                                    unset($v['resources']['enabled']);
-                                }
-                            }
-
-                            return $v;
-                        })
-                    ->end()
-                    ->addDefaultsIfNotSet()
-                    ->children()
-                        ->arrayNode('resources', 'resource')
-                            ->normalizeKeys(false)
-                            ->useAttributeAsKey('name')
-                            ->requiresAtLeastOneElement()
-                            ->acceptAndWrap(['string'], 'default')
-                            ->beforeNormalization()
-                                ->ifArray()
-                                ->then(static function ($v) {
-                                    if (!array_is_list($v)) {
-                                        return $v;
-                                    }
-
-                                    $resources = [];
-                                    foreach ($v as $resource) {
-                                        $resources[] = \is_array($resource) && isset($resource['name'])
-                                            ? [$resource['name'] => $resource['value']]
-                                            : ['default' => $resource]
-                                        ;
-                                    }
-
-                                    return array_merge_recursive([], ...$resources);
-                                })
-                            ->end()
-                            ->prototype('scalar')->end()
-                        ->end()
-                    ->end()
+                ->variableNode('semaphore')
+                    ->aliasOf('semaphore')
+                    ->treatFalseLike(['enabled' => false])
+                    ->treatTrueLike(['enabled' => true])
+                    ->beforeNormalization()->ifString()->then(static fn ($v) => ['resources' => $v])->end()
                 ->end()
             ->end()
         ;