+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -170,12 +170,6 @@ use Symfony\Component\Scheduler\Messenger\SchedulerTransportFactory;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
-use Symfony\Component\Semaphore\PersistingStoreInterface as SemaphoreStoreInterface;
-use Symfony\Component\Semaphore\Semaphore;
-use Symfony\Component\Semaphore\SemaphoreFactory;
-use Symfony\Component\Semaphore\Serializer\SemaphoreKeyNormalizer;
-use Symfony\Component\Semaphore\Store\LockStore;
-use Symfony\Component\Semaphore\Store\StoreFactory as SemaphoreStoreFactory;
 use Symfony\Component\Serializer\Attribute as SerializerMapping;
 use Symfony\Component\Serializer\Attribute\ExtendsSerializationFor;
 use Symfony\Component\Serializer\Encoder\DecoderInterface;
@@ -478,10 +472,6 @@ class FrameworkExtension extends Extension
 
         if ($this->readConfigEnabled('lock', $container, $config['lock'])) {
             $this->registerLockConfiguration($config['lock'], $container, $loader);
-        }
-
-        if ($this->readConfigEnabled('semaphore', $container, $config['semaphore'])) {
-            $this->registerSemaphoreConfiguration($config['semaphore'], $container, $loader);
         }
 
         if ($this->readConfigEnabled('rate_limiter', $container, $config['rate_limiter'])) {
@@ -2116,57 +2106,6 @@ class FrameworkExtension extends Extension
                 $container->setAlias(LockFactory::class, new Alias('lock.factory', false));
             } else {
                 $container->registerAliasForArgument('lock.'.$resourceName.'.factory', LockFactory::class, $resourceName.'.lock.factory', $resourceName);
-            }
-        }
-    }
-
-    private function registerSemaphoreConfiguration(array $config, ContainerBuilder $container, PhpFileLoader $loader): void
-    {
-        if (!class_exists(Semaphore::class)) {
-            throw new LogicException('Semaphore support cannot be enabled as the Semaphore component is not installed. Try running "composer require symfony/semaphore".');
-        }
-
-        $loader->load('semaphore.php');
-
-        // BC layer Semaphore < 7.4
-        if (!interface_exists(DenormalizerInterface::class) || !class_exists(SemaphoreKeyNormalizer::class)) {
-            $container->removeDefinition('serializer.normalizer.semaphore_key');
-        }
-
-        foreach ($config['resources'] as $resourceName => $resourceStore) {
-            $storeDsn = $container->resolveEnvPlaceholders($resourceStore, null, $usedEnvs);
-
-            if (str_starts_with($storeDsn, 'lock://') && !class_exists(LockStore::class)) {
-                throw new LogicException('Cannot use a lock store as the installed version of the Semaphore component does not support it. Try running "composer require symfony/semaphore:^8.1".');
-            }
-
-            $storeDefinition = new Definition(SemaphoreStoreInterface::class);
-            $storeDefinition->setFactory([SemaphoreStoreFactory::class, 'createStore']);
-            $storeDefinition->setArguments([match (true) {
-                $usedEnvs => $resourceStore,
-                str_starts_with($storeDsn, 'lock://') => new Reference('lock.'.(substr($storeDsn, 7) ?: 'default').'.factory'),
-                !str_contains($resourceStore, '://') => new Reference($resourceStore),
-                default => $resourceStore,
-            }]);
-
-            $container->setDefinition($storeDefinitionId = '.semaphore.'.$resourceName.'.store.'.$container->hash($storeDsn), $storeDefinition);
-
-            // Generate factories for each resource
-            $factoryDefinition = new ChildDefinition('semaphore.factory.abstract');
-            $factoryDefinition->replaceArgument(0, new Reference($storeDefinitionId));
-            $container->setDefinition('semaphore.'.$resourceName.'.factory', $factoryDefinition);
-
-            // Generate services for semaphore instances
-            $semaphoreDefinition = new Definition(Semaphore::class);
-            $semaphoreDefinition->setFactory([new Reference('semaphore.'.$resourceName.'.factory'), 'createSemaphore']);
-            $semaphoreDefinition->setArguments([$resourceName]);
-
-            // provide alias for default resource
-            if ('default' === $resourceName) {
-                $container->setAlias('semaphore.factory', new Alias('semaphore.'.$resourceName.'.factory', false));
-                $container->setAlias(SemaphoreFactory::class, new Alias('semaphore.factory', false));
-            } else {
-                $container->registerAliasForArgument('semaphore.'.$resourceName.'.factory', SemaphoreFactory::class, $resourceName.'.semaphore.factory', $resourceName);
             }
         }
     }