+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/semaphore_lock.php
@@ -1,6 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'lock' => 'flock',
-    'semaphore' => 'lock://',
-]);