+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/semaphore_service.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->register('my_service', \Redis::class);
-
-$container->loadFromExtension('framework', [
-    'semaphore' => 'my_service',
-]);