+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTestCase.php
@@ -106,7 +106,7 @@ use Symfony\Component\PropertyAccess\PropertyAccessor;
 use Symfony\Component\RemoteEvent\Messenger\ConsumeRemoteEventHandler;
 use Symfony\Component\RemoteEvent\RemoteEventBundle;
 use Symfony\Component\Security\Core\AuthenticationEvents;
-use Symfony\Component\Semaphore\Store\LockStore;
+use Symfony\Component\Semaphore\SemaphoreBundle;
 use Symfony\Component\Semaphore\Store\StoreFactory as SemaphoreStoreFactory;
 use Symfony\Component\Serializer\DependencyInjection\SerializerPass;
 use Symfony\Component\Serializer\Mapping\Loader\XmlFileLoader;
@@ -454,6 +454,25 @@ abstract class FrameworkExtensionTestCase extends TestCase
     {
         yield 'underscored' => ['legacy_remote_event'];
         yield 'hyphenated' => ['legacy_hyphenated_remote_event'];
+    }
+
+    public function testSemaphoreConfigurationIsForwardedToSemaphoreBundle()
+    {
+        $container = $this->createContainer(['kernel.charset' => 'UTF-8', 'kernel.secret' => 'secret', 'kernel.runtime_environment' => 'test']);
+        $container->registerExtension(new FrameworkExtension());
+        $container->registerExtension(new SemaphoreBundle()->getContainerExtension());
+        $this->loadFromFile($container, 'legacy_semaphore');
+        $container->getCompilerPassConfig()->setBeforeOptimizationPasses([]);
+        $container->getCompilerPassConfig()->setOptimizationPasses([]);
+        $container->getCompilerPassConfig()->setBeforeRemovingPasses([]);
+        $container->getCompilerPassConfig()->setRemovingPasses([]);
+        $container->getCompilerPassConfig()->setAfterRemovingPasses([]);
+        $container->compile();
+
+        $this->assertTrue($container->hasDefinition('semaphore.default.factory'));
+        $storeDef = $container->getDefinition($container->getDefinition('semaphore.default.factory')->getArgument(0));
+        $this->assertSame([SemaphoreStoreFactory::class, 'createStore'], $storeDef->getFactory());
+        $this->assertSame('redis://localhost', $storeDef->getArgument(0));
     }
 
     public function testEnabledPhpErrorsConfig()
@@ -3536,72 +3555,6 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $connection = $storeDef->getArgument(0);
         $this->assertInstanceOf(Reference::class, $connection);
         $this->assertEquals('my_service', $connection->__toString());
-    }
-
-    public function testDefaultSemaphore()
-    {
-        $container = $this->createContainerFromFile('semaphore');
-
-        $this->assertTrue($container->hasDefinition('semaphore.default.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.default.factory')->getArgument(0));
-        $this->assertSame('redis://localhost', $storeDef->getArgument(0));
-    }
-
-    public function testNamedSemaphores()
-    {
-        $container = $this->createContainerFromFile('semaphore_named');
-
-        $this->assertTrue($container->hasDefinition('semaphore.foo.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.foo.factory')->getArgument(0));
-        $this->assertSame('redis://paas.com', $storeDef->getArgument(0));
-
-        $this->assertTrue($container->hasDefinition('semaphore.qux.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.qux.factory')->getArgument(0));
-        $this->assertStringContainsString('REDIS_DSN', $storeDef->getArgument(0));
-    }
-
-    public function testSemaphoreWithService()
-    {
-        $container = $this->createContainerFromFile('semaphore_service', [], true, false);
-        $container->getCompilerPassConfig()->setOptimizationPasses([new ResolveChildDefinitionsPass()]);
-        $container->compile();
-
-        $this->assertTrue($container->hasDefinition('semaphore.default.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.default.factory')->getArgument(0));
-        $this->assertEquals(new Reference('my_service'), $storeDef->getArgument(0));
-    }
-
-    public function testSemaphoreWithLock()
-    {
-        if (!class_exists(LockStore::class)) {
-            $this->markTestSkipped('LockStore not available');
-        }
-
-        $container = $this->createContainerFromFile('semaphore_lock');
-
-        $this->assertTrue($container->hasDefinition('semaphore.default.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.default.factory')->getArgument(0));
-        $this->assertSame([SemaphoreStoreFactory::class, 'createStore'], $storeDef->getFactory());
-        $this->assertEquals(new Reference('lock.default.factory'), $storeDef->getArgument(0));
-    }
-
-    public function testSemaphoreWithNamedLock()
-    {
-        if (!class_exists(LockStore::class)) {
-            $this->markTestSkipped('LockStore not available');
-        }
-
-        $container = $this->createContainerFromFile('semaphore_lock_named');
-
-        $this->assertTrue($container->hasDefinition('semaphore.default.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.default.factory')->getArgument(0));
-        $this->assertSame([SemaphoreStoreFactory::class, 'createStore'], $storeDef->getFactory());
-        $this->assertEquals(new Reference('lock.default.factory'), $storeDef->getArgument(0));
-
-        $this->assertTrue($container->hasDefinition('semaphore.bar.factory'));
-        $storeDef = $container->getDefinition($container->getDefinition('semaphore.bar.factory')->getArgument(0));
-        $this->assertSame([SemaphoreStoreFactory::class, 'createStore'], $storeDef->getFactory());
-        $this->assertEquals(new Reference('lock.foo.factory'), $storeDef->getArgument(0));
     }
 
     public function testJsonStreamerEnabled()