+++ src/Symfony/Component/Semaphore/Tests/DependencyInjection/Fixtures/semaphore_named.php
@@ -1,0 +1,8 @@
+<?php
+
+$container->setParameter('env(REDIS_DSN)', 'redis://paas.com');
+
+$container->loadFromExtension('semaphore', [
+    'foo' => 'redis://paas.com',
+    'qux' => '%env(REDIS_DSN)%',
+]);